Dashboard component is now TicketList, moved authentication into separate directory
